Data_Ingestion_VDB/requirements.txt
# PDF Processing
PyMuPDF>=1.23.0

# Google Cloud for Vertex AI
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Vector Store - MILVUS instead of ChromaDB
pymilvus>=2.3.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0